Added the option to use a file's modification time for versioning
Setting the "mtime" config option to a true value will cause the plugin
to use each file's modification time instead of the application's
version in the versioned uri. An "include_path" config flag is also
added, that will specify the directories where a file may be found.  If
the file isn't found, we fallback to the use the application's version.

Each process of your Catalyst app will check and cache the mtime of each
matched uri once. This adds a bit of overhead, but allows more fine
grained control over cache busting. Using mtime may result in the same
file being served with different versions when running your app on
multiple servers. Using an MD5 hash of the file would be more precise,
but I do not believe it is worth the added overhead.